refactor(agent-farm): extract builder-lookup, reuse resolveAgentName
Move findBuilderById / findBuilderByIssue out of commands/attach.ts into
a shared lib/builder-lookup.ts module so afx dev (coming next) can reuse
the same implementation without duplicating the Tower-fallback path.

Migrate ID matching to `resolveAgentName` from @cluesmith/codev-core/
agent-names (lifted to core by #729 for exactly this kind of reuse):

- Before: `b.id.startsWith(id) || b.id.includes(id)` — substring inclusion,
  which silently matched IDs in the middle of unrelated builder names
- After: case-insensitive exact (with leading zeros stripped) → tail-match
  (e.g. "109" → "builder-spir-109"), matching VSCode's openBuilderByRoleOrId

This is a tightening, not a widening. Realistic inputs use tail-match
anyway; the old substring path was a too-permissive fallback that
resolveAgentName deliberately replaced.

Applied to both the local-state path and the Tower terminal_sessions
fallback path so attach and dev see identical resolution behavior
regardless of whether the builder lives in local state or only in Tower.

attach.ts now imports from lib/builder-lookup.js. All 28 attach tests
still pass; full vitest run 2594 passed / 13 skipped.

Refs #689.
